+++ src/runtime/Runtime.hpp
@@ -37,6 +37,7 @@ private:
     void performCustomMounts() const;
     void performDeviceMounts() const;
     void remountRootfsWithNoSuid() const;
+    void clearEnvironmentVariables() const;
 
 private:
     std::shared_ptr<common::Config> config;