+++ CHANGELOG.md
@@ -14,6 +14,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Added the `--digests` option to `sarus images` for displaying digests of locally available images
 - Added the `--username` and `--password-stdin` options to `sarus pull` for supplying authentication credentials directly on the command line.
   More details [here](https://sarus.readthedocs.io/en/stable/user/user_guide.html#pulling-images-from-private-repositories)
+- Added support for the optional `containersPolicy` parameter in the `sarus.json` configuration file.
+  More details [here](https://sarus.readthedocs.io/en/stable/config/configuration_reference.html#containerspolicy-object-optional).
+- Added support for the optional `containersRegistries.dPath` parameter in the `sarus.json` configuration file.
+  More details [here](https://sarus.readthedocs.io/en/stable/config/configuration_reference.html#containersregistries-dpath-string-optional).
 - Added support for labels defined in OCI image configurations
 
 ### Changed